requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi==0.115.12
uvicorn==0.34.2
psycopg2-binary==2.9.10
requests==2.32.3
sentence-transformers==4.1.0
